requirements.txt
Django==5.2.7
djangorestframework==3.16.1
django-cors-headers==4.9.0
psycopg2-binary==2.9.10
python-dotenv==1.0.1
whitenoise==6.7.0
gunicorn==22.0.0
dj-database-url==2.1.0
drf-yasg==1.21.8
django-filter==23.5
celery==5.3.6
redis==5.0.1
pytest==7.4.3
pytest-django==4.8.0
pytest-cov==4.1.0
reportlab==4.1.0
openpyxl==3.1.2